requirements.txt
crewai
crewai_tools
python_dotenv
openai

